New "enstackener" command.
You put a bunch of stuff on the stack, and a number that indicates how
many items you want to be in the new enstackener word, execute the
coomand, and it creates a new command with that many of the items from
the stack in it that, when executed, will put them onto the stack.

Got that?