Login/out and DB goodness.
Created initialize_database.py script to set up the Sqlite DB (and if you
pass it any extra arg it will also wipe the DB before re-initializing
it.)

Made a simple login system. N.B.: we should "salt" the passwords or
something.  I'm sure there's an off-the-shelf component to handle this
and I'll go find it soon, but for now we can create users (see
http://pythonhosted.org/Flask-SQLAlchemy/quickstart.html for details) and
login using the accounts.