OpenID works, and will create new users. ^_^
I modified the initialize_database.py script to add a dummy email for Ed
and an extra step to protect the DB from accidentally dropping the
tables.

Added and "email" field to the User class which is automatically
populated from the OpenID info if sign-in/up succeeds.

New users created by the OpenID login WILL NOT HAVE their passwords set,
and there is currently no way to later set a password, we should fix that
eventually.